refactor: 기존 NewCard~ 를 UpdateCard로 변경
