+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.Payments"
         tools:targetApi="31">
         <activity
-            android:name=".screens.card.new.NewCardActivity"
+            android:name=".screens.card.update.UpdateCardActivity"
             android:exported="false">
         </activity>
