+++ app/src/main/res/values/strings.xml
@@ -5,16 +5,18 @@
     <string name="all_back">뒤로 가기</string>
     <string name="all_done">완료</string>
 
-    <!-- NewCard -->
-    <string name="new_card_top_bar_title">카드 추가</string>
-    <string name="new_card_card_number_label">카드 번호</string>
-    <string name="new_card_card_number_placeholder">0000 - 0000 - 0000 - 0000</string>
-    <string name="new_card_expiration_day_label">만료일</string>
-    <string name="new_card_expiration_day_placeholder">MM / YY</string>
-    <string name="new_card_card_owner_name_label">카드 소유자 이름(선택)</string>
-    <string name="new_card_card_owner_name_placeholder">카드에 표시된 이름을 입력하세요.</string>
-    <string name="new_card_password_label">비밀번호</string>
-    <string name="new_card_password_placeholder">0000</string>
+    <!-- UpdateCard -->
+    <string name="add_card_top_bar_title">카드 추가</string>
+    <string name="update_card_card_number_label">카드 번호</string>
+    <string name="update_card_card_number_placeholder">0000 - 0000 - 0000 - 0000</string>
+    <string name="update_card_expiration_day_label">만료일</string>
+    <string name="update_card_expiration_day_placeholder">MM / YY</string>
+    <string name="update_card_card_owner_name_label">카드 소유자 이름(선택)</string>
+    <string name="update_card_card_owner_name_placeholder">카드에 표시된 이름을 입력하세요.</string>
+    <string name="update_card_password_label">비밀번호</string>
+    <string name="update_card_password_placeholder">0000</string>
+    <string name="edit_card_top_bar_title">카드 수정</string>
+
 
     <!-- CardList -->
     <string name="card_list_title">Payments</string>
